agent/requirements.txt
boto3==1.37.5
pyOpenSSL==25.1.0
fastapi==0.115.13
uvicorn==0.34.3